refactor: Change variable name in dynamic component loading
Change the variable name from 'm' to 'c' in the dynamic component loading for the LoginComponent. Initially, 'm' stood for 'module' as it was a module for the login page. However, as it's now a component, 'c' is more appropriate. This resolves the issue raised by @boyzwhocried.

File affected: AppRoutingModule